refactor: 유틸성 패키지 위치 이동
