+++ src/main/java/kr/wootecat/dongle/model/http/session/RandomIdGenerator.java
@@ -1,4 +1,4 @@
-package kr.wootecat.dongle.utils;
+package kr.wootecat.dongle.model.http.session;
 
 import java.util.UUID;
