+++ src/main/java/kr/wootecat/dongle/utils/IOUtils.java
@@ -1,4 +1,4 @@
-package utils;
+package kr.wootecat.dongle.utils;
 
 import java.io.BufferedReader;
 import java.io.IOException;
