+++ src/main/java/kr/wootecat/dongle/utils/SplitUtils.java
@@ -1,4 +1,4 @@
-package utils;
+package kr.wootecat.dongle.utils;
 
 import java.util.HashMap;
 import java.util.Map;
