+++ src/test/java/kr/wootecat/dongle/utils/RandomIdGeneratorTest.java
@@ -5,6 +5,9 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import kr.wootecat.dongle.model.http.session.IdGenerator;
+import kr.wootecat.dongle.model.http.session.RandomIdGenerator;
+
 class RandomIdGeneratorTest {
 
     @DisplayName("랜덤 ID 생성기를 통해 생성한 값이 문자열이다. 값은 랜덤값으로 생성되기 때문에 타입만 확인한다.")
