+++ src/test/java/utils/FileIoUtilsTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import kr.wootecat.dongle.utils.FileIoUtils;
+
 public class FileIoUtilsTest {
     private static final Logger log = LoggerFactory.getLogger(FileIoUtilsTest.class);
